[MicroBlaze Implementations] - Changed name of interface IlpScenario to Mapper;- Changed MbElfReader and MbTraceReader to conform to the new interface; - MbInstructionBlockWriter now receives the baseFilename, instead of the complete trace filename; - Removed some output messages from certain transformations; - Added method getFilesRecursive applied to every file in a folder, instead of only files with certain extensions;

+++ src/microblaze/org/ancora/MicroBlaze/Trace/TraceProperties.java
@@ -52,14 +52,14 @@ public class TraceProperties {
       return Float.parseFloat(value);
    }
 
-   public int getInstructions() {
+   public long getInstructions() {
      String value = TraceUtils.safeGet(Key.instructions, properties);
-      return Integer.parseInt(value);
+      return Long.parseLong(value);
    }
 
-   public int getCycles() {
+   public long getCycles() {
       String value = TraceUtils.safeGet(Key.cycles, properties);
-      return Integer.parseInt(value);
+      return Long.parseLong(value);
    }
 
    public enum Key {
